requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
click==8.2.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_enquetes==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
